requirements.txt
azure-core==1.28.0
azure-identity==1.13.0
azure-storage-queue==12.6.0